+++ src/Language/Mimsa/Types.hs
@@ -17,6 +17,8 @@ module Language.Mimsa.Types
     module Language.Mimsa.Types.Printer,
     module Language.Mimsa.Types.ResolverError,
     module Language.Mimsa.Types.InterpreterError,
+    module Language.Mimsa.Types.Variable,
+    module Language.Mimsa.Types.Scope,
   )
 where
 
@@ -29,8 +31,10 @@ import Language.Mimsa.Types.Name
 import Language.Mimsa.Types.Printer
 import Language.Mimsa.Types.ResolverError
 import Language.Mimsa.Types.Scheme
+import Language.Mimsa.Types.Scope
 import Language.Mimsa.Types.Store
 import Language.Mimsa.Types.TypeError
 import Language.Mimsa.Types.Typechecker
+import Language.Mimsa.Types.Variable
 
 ------
