+++ src/Language/Mimsa/Types/Variable.hs
@@ -1,0 +1,16 @@
+module Language.Mimsa.Types.Variable where
+
+import qualified Data.Text as T
+import Language.Mimsa.Types.Name
+import Language.Mimsa.Types.Printer
+
+data Variable
+  = NamedVar Name
+  | NumberedVar Int
+  | BuiltIn Name
+  deriving (Eq, Ord, Show)
+
+instance Printer Variable where
+  prettyPrint (NamedVar n) = prettyPrint n
+  prettyPrint (NumberedVar i) = T.pack (show i)
+  prettyPrint (BuiltIn n) = prettyPrint n
